Root/PPT_Agent/archive_v1/requirements.txt
# Core
google-generativeai
python-pptx
pdf2image
pymupdf
pydantic
python-dotenv

# CLI/UI
typer
rich

# Utils
pillow
numpy
